requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
altair==6.0.0
amplitude-analytics==1.2.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
apprise==1.9.8
asgi-lifespan==2.1.0
async-timeout==5.0.1
asyncpg==0.31.0
asyncssh==2.22.0
atpublic==7.0.0
attrs==25.4.0
beartype==0.22.9
billiard==4.2.4
blinker==1.9.0
cachetools==7.0.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
contourpy==1.3.2
coolname==4.0.0
cronsim==2.7
cryptography==46.0.5
cycler==0.12.1
cyclopts==4.8.0
databricks-sdk==0.98.0
dateparser==1.3.0
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.4
dpath==2.2.0
dulwich==1.1.0
dvc==3.66.1
dvc-data==3.18.3
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.1
fakeredis==2.34.1
fastapi==0.135.1
filelock==3.25.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
gto==1.9.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
humanfriendly==10.0
humanize==4.15.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.11
imbalanced-learn==0.14.1
importlib_metadata==8.7.1
iniconfig==2.3.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kombu==5.6.2
lupa==2.6
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
mpmath==1.3.0
multidict==6.7.1
narwhals==2.17.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.23.2
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pendulum==3.2.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
pluggy==1.6.0
prefect==3.6.21
prettytable==3.17.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
py-key-value-aio==0.4.4
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pydocket==0.18.2
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.3.2
pyreadline3==3.5.4
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-slugify==8.0.4
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
readchar==4.2.1
redis==7.3.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
ruff==0.15.8
scikit-learn==1.7.2
scipy==1.15.3
scmrepo==3.6.1
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
skl2onnx==1.20.0
sklearn-compat==0.1.5
skops==0.13.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.48
sqlparse==0.5.5
sqltrie==0.11.2
starlette==0.52.1
streamlit==1.55.0
sympy==1.14.0
tabulate==0.10.0
taskgroup==0.2.2
tenacity==9.1.4
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.4.0
tomlkit==0.14.0
tornado==6.5.4
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uncalled-for==0.2.0
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
voluptuous==0.16.0
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.6
yarl==1.23.0
zc.lockfile==4.0
zipp==3.23.0
